backend/fastap/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
rembg==2.0.59
pillow==10.4.0
python-multipart
boto3
